+++ src/StateModifier.cpp
@@ -148,16 +148,16 @@ void StateModifier::setApplyAndRecord(const StateChange& stateChange) {
 		m_configManager.configParams().onKeyPressed(std::get<SDL_Scancode>(stateChange.value), *this);
 		break;
 	case StateChangeType::Param_Wheel:
-		m_configManager.configParams().setApplyAndRecord_Wheel(std::get<float>(stateChange.value), *this);
+		m_configManager.configParams().setApplyAndRecord_Wheel(std::get<int>(stateChange.value), *this);
 		break;
 	case StateChangeType::Param_CtrlWheel:
-		m_configManager.configParams().setApplyAndRecord_CtrlWheel(std::get<float>(stateChange.value), *this);
+		m_configManager.configParams().setApplyAndRecord_CtrlWheel(std::get<int>(stateChange.value), *this);
 		break;
 	case StateChangeType::Param_ShiftWheel:
-		m_configManager.configParams().setApplyAndRecord_ShiftWheel(std::get<float>(stateChange.value), *this);
+		m_configManager.configParams().setApplyAndRecord_ShiftWheel(std::get<int>(stateChange.value), *this);
 		break;
 	case StateChangeType::Param_AltWheel:
-		m_configManager.configParams().setApplyAndRecord_AltWheel(std::get<float>(stateChange.value), *this);
+		m_configManager.configParams().setApplyAndRecord_AltWheel(std::get<int>(stateChange.value), *this);
 		break;
 	case StateChangeType::Mouse_ForceField:
 		m_mouseInteractions.setForceField(std::get<glm::vec2>(stateChange.value));
